feat(bookmarks): add clipboard context-menu
Add logic to copy the URL of a bookmark to the clipboard.
Display information toast.
