+++ .air.toml
@@ -18,27 +18,27 @@ tmp_dir = "./tmp"
   # -------------------------------------------------------------------------
   # bookmarks
   # -------------------------------------------------------------------------
-  #args_bin = [
-  #  "--port",
-  #  "3003",
-  #  "--basepath",
-  #  "./internal/bookmarks"
-  #]
-  #bin = "./tmp/bookmarks.api"
-  #cmd = "go build -ldflags=\"-X main.Version=NOW -X main.Build=DEV\" -o ./tmp/bookmarks.api ./cmd/bookmarks/server/main.go"
+  args_bin = [
+    "--port",
+    "3003",
+    "--basepath",
+    "./internal/bookmarks"
+  ]
+  bin = "./tmp/bookmarks.api"
+  cmd = "go build -ldflags=\"-X main.Version=NOW -X main.Build=DEV\" -o ./tmp/bookmarks.api ./cmd/bookmarks/server/main.go"
 
 
   # -------------------------------------------------------------------------
   # mydms
   # -------------------------------------------------------------------------
-  args_bin = [
-    "--port",
-    "3003",
-    "--basepath",
-    "./internal/mydms"
-  ]
-  bin = "./tmp/mydms.api"
-  cmd = "go build -ldflags=\"-X main.Version=NOW -X main.Build=DEV\" -o ./tmp/mydms.api ./cmd/mydms/server/main.go"
+  #args_bin = [
+  #  "--port",
+  #  "3003",
+  #  "--basepath",
+  #  "./internal/mydms"
+  #]
+  #bin = "./tmp/mydms.api"
+  #cmd = "go build -ldflags=\"-X main.Version=NOW -X main.Build=DEV\" -o ./tmp/mydms.api ./cmd/mydms/server/main.go"
 
   delay = 500
   exclude_dir = ["assets", "tmp", "vendor", "testdata", "frontend", "caddy", ".vscode", "doc", "tmp", "tools", "upload", "litestream", "_logs"]
