+++ compose-integration.yaml
@@ -12,8 +12,8 @@ services:
                 image: litestream
                 user: "1000:1000"
                 volumes:
-                        - ./litestream/litestream.yml:/etc/litestream.yml
-                        - ./testdata/sqlite/integration:/opt/litestream/store
+                        - ./litestream/litestream.yml:/etc/litestream.yml:z
+                        - ./testdata/sqlite/integration:/opt/litestream/store:z
                 environment:
                         LITESTREAM_ACCESS_KEY_ID: $LITESTREAM_ACCESS_KEY_ID
                         LITESTREAM_SECRET_ACCESS_KEY: $LITESTREAM_SECRET_ACCESS_KEY
@@ -28,7 +28,7 @@ services:
         minio:
                 image: minio/minio
                 volumes:
-                        - ./testdata/minio:/data
+                        - ./testdata/minio:/data:z
                 environment:
                         - "MINIO_ROOT_USER=s3_access_key"
                         - "MINIO_ROOT_PASSWORD=s3_access_secret"
@@ -95,10 +95,10 @@ services:
                 image: core
                 user: "1000:1000"
                 volumes:
-                        - ./internal/core:/opt/core/etc
-                        - ./_logs:/opt/core/logs
-                        - ./upload:/opt/core/uploads
-                        - ./testdata/sqlite/integration:/store
+                        - ./internal/core:/opt/core/etc:z
+                        - ./_logs:/opt/core/logs:z
+                        - ./upload:/opt/core/uploads:z
+                        - ./testdata/sqlite/integration:/store:z
                 restart: always
                 environment:
                         CO_BASECONFIG__ENVIRONMENT: Integration
@@ -129,10 +129,10 @@ services:
                 image: mydms
                 user: "1000:1000"
                 volumes:
-                        - ./internal/mydms:/opt/mydms/etc
-                        - ./_logs:/opt/mydms/logs
-                        - ./testdata/sqlite/integration:/store
-                        - /tmp:/opt/mydms/uploads
+                        - ./internal/mydms:/opt/mydms/etc:z
+                        - ./_logs:/opt/mydms/logs:z
+                        - ./testdata/sqlite/integration:/store:z
+                        - ./upload:/opt/mydms/uploads:z
                 restart: always
                 environment:
                         MY_BASECONFIG__ENVIRONMENT: Integration
@@ -165,9 +165,9 @@ services:
                 image: bookmarks
                 user: "1000:1000"
                 volumes:
-                        - ./internal/bookmarks:/opt/bookmarks/etc
-                        - ./_logs:/opt/bookmarks/logs
-                        - ./testdata/sqlite/integration:/store
+                        - ./internal/bookmarks:/opt/bookmarks/etc:z
+                        - ./_logs:/opt/bookmarks/logs:z
+                        - ./testdata/sqlite/integration:/store:z
                 restart: always
                 environment:
                         BM_BASECONFIG__ENVIRONMENT: Integration
@@ -194,6 +194,6 @@ services:
                 ports:
                         - "443:443"
                 volumes:
-                        - ./caddy/Caddyfile:/etc/caddy/Caddyfile
-                        - ./caddy/certs:/opt/caddy/certs
+                        - ./caddy/Caddyfile:/etc/caddy/Caddyfile:z
+                        - ./caddy/certs:/opt/caddy/certs:z
                 restart: on-failure
