+++ pkg/handler/html/fragment_toast.html
@@ -18,5 +18,14 @@
             </div>
             <div class="toast-body" id="toast_messsage_text-error"></div>
         </div>
+        <div id="toastMessage-info" class="toast" role="alert" data-type="info">
+            <div class="toast-header" style="background-color: var(--bs-info);color:white;">
+                <i class="bi bi-info-circle"></i>&nbsp;
+                <strong class="me-auto" id="toast_message_title-info"></strong>
+                <small></small>
+                <button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
+            </div>
+            <div class="toast-body" id="toast_messsage_text-info"></div>
+        </div>
     </div>
 </div>
